requirements.txt
Django==4.2.10
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.1.0
django-filter==23.5
django-tables2==2.6.0
xlwt==1.3.0
reportlab==4.0.9
django-import-export==3.3.1